requirements.txt
Flask
pymongo
numpy
tensorflow
Pillow
pytesseract

